Truncation error not recorded during two-site DMRG
The truncation error was not recorded during the two-site DMRG because the variable result_array in DMRG.finite_sweep() is initialized with np.nan. Later on, when the truncation error should be added to this variable, the result is np.nan again. Corrected this by initializing the portion in result_array reserved for the truncation error with zeros rather than np.nan if two-site DMRG is used.